deviceHandler: rename file
